% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_tables.R
\name{add_missing_table_headers}
\alias{add_missing_table_headers}
\title{Add Missing Table Headers}
\usage{
add_missing_table_headers(text_string)
}
\arguments{
\item{text_string}{The string to check against the pattern and replace as
needed.}
}
\value{
Either the original \code{text_string} or one prepended with
"Performance Metric".
}
\description{
Sometimes, the table row containing the fiscal year headers (FY 2018,
FY 2019...) does not start with the words "Performance Metric", which leaves
the first column of the processed text empty. We want that column to say
"Performance Metric", so we will add it.
}
