% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_tables.R
\name{brute_force_row_fix}
\alias{brute_force_row_fix}
\title{Fix Edge Cases}
\usage{
brute_force_row_fix(text_string)
}
\arguments{
\item{text_string}{The string to process. Only the first element is
processed.}
}
\value{
a list including the trimmed string, the table title, and all other
elements that were part of the \code{text_string} passed to the function.
}
\description{
When the more general patterns fail, use more specific patterns. This
function cleans up most of the remaining mess the others leave behind.
}
