% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_excel.R
\name{build_pivot}
\alias{build_pivot}
\title{Build Pivot Table}
\usage{
build_pivot(
  data,
  metric_type_filter = c("integer", "double", "percent", "text")
)
}
\arguments{
\item{data}{A table of MDUFA metrics, such as \code{mdufa::mdufa4}.}

\item{metric_type_filter}{The metric type to filter to before pivoting, equal
to either \code{"integer"}, \code{"percent"}, \code{"double"}, or
\code{"text"}.}
}
\value{
A tibble.
}
\description{
Pivot data by year, filtered to a metric type
}
