% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_report.R
\name{build_submission_type_regex}
\alias{build_submission_type_regex}
\title{Build Submission Type Regex}
\usage{
build_submission_type_regex(mdufa_period)
}
\arguments{
\item{mdufa_period}{The MDUFA period}
}
\value{
A regex pattern
}
\description{
Creates a regex pattern for matching submission types based on MDUFA period.
}
\keyword{internal}
