% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{color_fda_blue}
\alias{color_fda_blue}
\title{FDA Blue Color}
\format{
Character string with hex color code
}
\usage{
color_fda_blue
}
\description{
The primary blue color used in FDA visualizations and branding.
}
\keyword{datasets}
