% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{comparable_ohts}
\alias{comparable_ohts}
\title{Comparable OHT Organizations}
\format{
Character vector of OHT codes
}
\usage{
comparable_ohts
}
\description{
OHT codes that can be compared across MDUFA III, IV, and V periods.
OHT1, OHT3, and OHT4 are excluded because they don't have clean
mappings from MDUFA III divisions.
}
\keyword{datasets}
