% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{convert_doubles}
\alias{convert_doubles}
\title{Convert Double Metrics}
\usage{
convert_doubles(data)
}
\arguments{
\item{data}{A MDUFA dataset, typically filtered to metric_type == "double"}
}
\value{
Dataset with value as double and fy as integer
}
\description{
Filters a MDUFA dataset to double metrics and converts the value column
to double/numeric type. Also converts fy to integer.
}
\examples{
library(dplyr)
mdufa4 |>
  filter(metric_type == "double") |>
  convert_doubles() |>
  head()
}
