% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{convert_percents}
\alias{convert_percents}
\title{Convert Percent Metrics}
\usage{
convert_percents(data)
}
\arguments{
\item{data}{A MDUFA dataset, typically filtered to metric_type == "percent"}
}
\value{
Dataset with value as proportion (0-1) and fy as integer
}
\description{
Filters a MDUFA dataset to percent metrics and converts the value column
to a proportion (0-1 scale). Also converts fy to integer.
}
\examples{
library(dplyr)
mdufa4 |>
  filter(metric_type == "percent") |>
  convert_percents() |>
  head()
}
