% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{date_phe_end}
\alias{date_phe_end}
\title{COVID-19 Public Health Emergency End Date}
\format{
Date object
}
\usage{
date_phe_end
}
\description{
The date the COVID-19 Public Health Emergency ended (May 11, 2023).
This date marks the end of the PHE period.
}
\keyword{datasets}
