% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{date_phe_start}
\alias{date_phe_start}
\title{COVID-19 Public Health Emergency Start Date}
\format{
Date object
}
\usage{
date_phe_start
}
\description{
The date the COVID-19 Public Health Emergency was declared
(January 27, 2020). This date marks the beginning of the PHE period
which affected FDA review timelines.
}
\keyword{datasets}
