% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_excel.R
\name{export_excel}
\alias{export_excel}
\title{Export to Excel}
\usage{
export_excel(data, filepath)
}
\arguments{
\item{data}{The data to export. Must have identical column names to
\code{colnames(mdufa::quarterly_performance)}.}

\item{filepath}{The \code{.xlsx} file to write to. Existing files with the
same name will not be overwritten.}
}
\value{
Nothing.
}
\description{
Export to Excel
}
\examples{
\dontrun{
filepath <- tempfile(fileext = ".xlsx")
export_excel(mdufa::quarterly_performance, filepath)
}
}
