% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_text.R
\name{extract_text}
\alias{extract_text}
\title{Extract Text from PDF Report}
\usage{
extract_text(url_report)
}
\arguments{
\item{url_report}{The URL of the report to read}
}
\value{
The embedded text of the report
}
\description{
Extract Text from PDF Report
}
