% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_text_m5.R
\name{extract_text_m5}
\alias{extract_text_m5}
\title{Extract Text from PDF MDUFA 5 Report}
\usage{
extract_text_m5(url_report)
}
\arguments{
\item{url_report}{The URL of the report to read}
}
\value{
A tibble that includes page numbers and the embedded text of the
report
}
\description{
These reports don't always have page numbers, complicating the procedure.
}
