% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_mdufa.R
\name{filter_comparable_ohts}
\alias{filter_comparable_ohts}
\title{Filter to Comparable OHT Organizations}
\usage{
filter_comparable_ohts(data)
}
\arguments{
\item{data}{Combined MDUFA dataset from \code{\link[=combine_mdufa_oht]{combine_mdufa_oht()}}}
}
\value{
Filtered dataset with only comparable OHTs
}
\description{
Filters a combined MDUFA dataset to only include OHT organizations
that can be compared across all three MDUFA periods (III, IV, V).
This excludes OHT1, OHT3, and OHT4 which don't have clean mappings
from MDUFA III divisions.
}
\details{
Comparable OHTs:
\itemize{
\item OHT2 (Cardiovascular) <- DCD
\item OHT5 (Neurological) <- DNPMD
\item OHT6 (Orthopedic) <- DOD
\item OHT7 (IVD) <- DCTD + DIHD + DMD + DMGP (aggregated)
\item OHT8 (Radiological) <- DRH
}
}
\examples{
combined <- combine_mdufa_oht()
comparable <- filter_comparable_ohts(combined)
unique(comparable$organization)
}
