% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_metrics.R
\name{filter_metrics}
\alias{filter_metrics}
\title{Filter Metrics}
\usage{
filter_metrics(data, metric_type)
}
\arguments{
\item{data}{A MDUFA performance table, such as
\code{mdufa::quarterly_performance}.}

\item{metric_type}{The metric type to filter the table to. Valid options
include \code{"integer"}, \code{"percent"}, \code{"double"}, and
\code{"text"}.}
}
\value{
A filtered performance table, with the \code{value} column in the
proper format.
}
\description{
Filter a MDUFA metrics table to one metric type, such as all metrics that are
a percent.
}
\examples{
mdufa::mdufa4 |>
  filter_metrics(metric_type = "percent")
}
