% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_empty_rows.R
\name{fix_empty_rows}
\alias{fix_empty_rows}
\title{Fix Empty Rows}
\usage{
fix_empty_rows(data, name_column)
}
\arguments{
\item{data}{A tibble}

\item{name_column}{The name of the column that functions as the row names and
may break across multiple rows}
}
\value{
A compressed tibble without empty rows and with repaired names
}
\description{
When converting a table from PDF, sometimes the \code{name_column} breaks
across multiple lines. Each of these lines can then appear as a separate
empty row in the resulting table. This function fixes that by checking for
empty rows and then concatenating the \code{name_column} value of the
following row with the \code{name_column} value of the previous row.
}
