% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_tables.R
\name{fix_goal_row}
\alias{fix_goal_row}
\title{Fix Goal Row}
\usage{
fix_goal_row(text_string)
}
\arguments{
\item{text_string}{The string to process.}
}
\value{
a list including the trimmed string, the table title, and all other
elements that were part of the \code{text_string} passed to the function.
}
\description{
Some of the metrics tables include performance goals, like "90\% within 180
days," or text along those lines. Between line break and delimiter issues,
these rows cause problems. This function attempts to fix those problems.
}
\details{
This function is also an example of what happens when one attempts to write
comments and documentation a long time after the code is written. For shame.
}
