% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_tables.R
\name{fix_si_row}
\alias{fix_si_row}
\title{Fix Substantive Interaction Rows}
\usage{
fix_si_row(text_string)
}
\arguments{
\item{text_string}{The string to process.}
}
\value{
The string with extraneous line breaks in the substantive interaction
goals removed.
}
\description{
This function fixes performance goals for substantive interactions that break
across lines. This is an issue in the MDUFA III reports.
}
\examples{
testthat::expect_equal(
  object =
    fix_si_row(
      # This text has extraneous line breaks
      paste0(
        "\nAverage number of FDA days to\nSubstantive Interaction|",
        "\nMaximum FDA days to\nSubstantive Interaction|",
        "\n40th Percentile FDA days to\nSubstantive Interaction|",
        "\n20th Percentile\nFDA\ndays to Substantive Interaction|",
        sep = "\n"
      )
    ),
  expected =
    # This text does not have extraneous line breaks
    paste0(
      "\nAverage number of FDA days to Substantive Interaction|",
      "\nMaximum FDA days to Substantive Interaction|",
      "\n40th Percentile FDA days to Substantive Interaction|",
      "\n20th Percentile FDA days to Substantive Interaction|",
      sep = "\n"
    )
)
}
