% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regex_patterns.R
\name{fy_start_line_pattern}
\alias{fy_start_line_pattern}
\title{Generate a regular expression pattern for table header rows with FYs in them}
\usage{
fy_start_line_pattern()
}
\value{
A regular expression pattern
}
\description{
The pattern matches a line that includes the string "FY" followed by a space
and four digits, repeated five times.
}
