% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_m5.R
\name{get_m5}
\alias{get_m5}
\title{Get MDUFA V Data}
\usage{
get_m5(report_description, report_link, report_date)
}
\arguments{
\item{report_description}{A description of the report from the FDA website.
Typically something along the lines of 'March 1, 2023 MDUFA V Performance
Report' or similar.}

\item{report_link}{The url at which the report PDF can be accessed.}

\item{report_date}{The date of the report.}
}
\value{
A table of MDUFA V data.
}
\description{
Get MDUFA V Data
}
