% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_tables.R
\name{get_page_number}
\alias{get_page_number}
\title{Get Page Number}
\usage{
get_page_number(text_string)
}
\arguments{
\item{text_string}{The string to process. Only the first element is
processed.}
}
\value{
a list including the trimmed string (not including the page number),
the page number, and all other elements that were part of the
\code{text_string} passed to the function.
}
\description{
Extract the page number from the first element of \code{text_string}, then
return the trimmed text string (without the page number) and all other
elements of the object, appending the page number.
}
