% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_tables.R
\name{get_table}
\alias{get_table}
\title{Get Table}
\usage{
get_table(text_string)
}
\arguments{
\item{text_string}{The string to process. Only the first element is
processed.}
}
\value{
a list including the current table as the first element followed by
all other elements that were part of the \code{text_string} passed to the
function.
}
\description{
Get Table
}
