% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_tables.R
\name{get_table_title}
\alias{get_table_title}
\title{Get Table Title}
\usage{
get_table_title(text_string)
}
\arguments{
\item{text_string}{The string to process. Only the first element is
processed.}
}
\value{
a list including the trimmed string (not including the table title),
the table title, and all other elements that were part of the
\code{text_string} passed to the function.
}
\description{
Extract the table title from the first element of \code{text_string}, then
return the trimmed text string (without the table title) and all other
elements of the object, appending the table title.
}
