% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regex_patterns.R
\name{goal_days_line_pattern}
\alias{goal_days_line_pattern}
\title{Generate a regular expression pattern for lines of goal days in a table}
\usage{
goal_days_line_pattern()
}
\value{
A regular expression pattern
}
\description{
This function generates a regular expression pattern for matching lines of
goal days in a table. The pattern matches a line that starts with a new line,
white space, and five repetitions of a pattern for goal days (delimited by
vertical bars), followed by white space and a new line.
}
