% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regex_patterns.R
\name{goal_days_pattern}
\alias{goal_days_pattern}
\title{Generate a regular expression pattern for goal days}
\usage{
goal_days_pattern()
}
\value{
A regular expression pattern
}
\description{
This function generates a regular expression pattern for matching goal days
in text. The pattern matches a string that contains zero or more digits,
followed by the word "FDA" or the word "Industry", followed by the word
"Days".
}
