% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regex_patterns.R
\name{goal_percent_line_pattern}
\alias{goal_percent_line_pattern}
\title{Goal Percent Line Pattern}
\usage{
goal_percent_line_pattern()
}
\value{
A regex pattern that matches to a line that includes
\code{goal_percent_pattern()}.
}
\description{
Goal Percent Line Pattern
}
