% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_tables.R
\name{insert_delim}
\alias{insert_delim}
\title{Insert Delimiter}
\usage{
insert_delim(text_string, whitespace = 2)
}
\arguments{
\item{text_string}{The string to process and insert delimiters into as
needed.}

\item{whitespace}{Defaults to \code{2}. The number of white-spaces to look
for and replace with the delimiter (\code{|}).}
}
\value{
The \code{text_string} with delimiters inserted as applicable.
}
\description{
Places the pipe character (\code{|}) between table columns, which are
assumed to be separated by at least \code{whitespace} number of space
characters.
}
\details{
This function only operates on the first element of \code{text_string}.
}
