% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{ivd_divisions}
\alias{ivd_divisions}
\title{IVD Division Identifiers}
\format{
Character vector of division codes
}
\usage{
ivd_divisions
}
\description{
MDUFA III division codes that map to OHT7 (In Vitro Diagnostics).
These four divisions were consolidated into OHT7 during the FDA
reorganization on September 30, 2019.
}
\keyword{datasets}
