% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{ivd_divisions_full}
\alias{ivd_divisions_full}
\title{IVD Division Full Names}
\format{
Character vector of full division names
}
\usage{
ivd_divisions_full
}
\description{
Full names for IVD divisions as they appear in MDUFA III reports.
}
\keyword{datasets}
