% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdufa2.R
\docType{data}
\name{mdufa2}
\alias{mdufa2}
\title{MDUFA II}
\format{
A tibble with 335 rows and 14 fields:

\describe{
\item{report_description}{chr "January 29, 2014 FDA MDUFA II Performance
Rep…}
\item{report_link}{chr "https://www.fda.gov/media/88098/download", "h…}
\item{report_date}{date 2014-01-29, 2014-01-29, 2014-01-29, 2014-01-2…}
\item{report_mdufa_period}{chr "MDUFA II", "MDUFA II", "MDUFA II", "MDUFA
II"…}
\item{source}{chr "MDUFA II Quarterly (Non expedited PMA Orginal…}
\item{page}{chr "1", "1", "1", "1", "1", "1", "1", "1", "1", "…}
\item{table_number}{chr "1", "1", "1", "1", "1", "1", "1", "1", "1", "…}
\item{organization}{chr "CDRH", "CDRH", "CDRH", "CDRH", "CDRH", "CDRH"…}
\item{program}{chr "PMA Original and Panel-track Supplements", "P…}
\item{table_title}{chr "Non expedited PMA Orginal and Panel-track Sup…}
\item{metric_type}{chr "integer", "integer", "integer", "integer", "i…}
\item{performance_metric}{chr "Workload (Filed to Date)", "Workload (Filed
t…}
\item{fy}{chr "2008", "2009", "2010", "2011", "2012", "2008"…}
\item{value}{chr "33", "39", "53", "44", "29", "33", "39", "53"…}
}
}
\source{
\href{https://www.fda.gov/industry/medical-device-user-fee-amendments-mdufa/mdufa-reports}{FDA MDUFA Reports} # nolint: line_length_linter.
accessed 2023-01-21.
}
\usage{
mdufa2
}
\description{
FDA's Quarterly Medical Device User Performance Metrics in a tidy data
format.
}
\keyword{datasets}
