% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdufa3.R
\docType{data}
\name{mdufa3}
\alias{mdufa3}
\title{MDUFA III}
\format{
A tibble with 7715 rows and 14 fields:

\describe{
\item{report_description}{chr "December 10, 2018 MDUFA III Performance
Repor…}
\item{report_link}{chr "https://public4.pagefreezer.com/content/17013…}
\item{report_date}{date 2018-12-10, 2018-12-10, 2018-12-10, 2018-12-1…}
\item{report_mdufa_period}{chr "MDUFA III", "MDUFA III", "MDUFA III", "MDUFA
…}
\item{source}{chr "Table 1.1. CDRH – PMA Original and Panel Trac…}
\item{page}{chr "19", "19", "19", "19", "19", "19", "19", "19"…}
\item{table_number}{chr "1.1", "1.1", "1.1", "1.1", "1.1", "1.1", "1.1…}
\item{organization}{chr "CDRH", "CDRH", "CDRH", "CDRH", "CDRH", "CDRH"…}
\item{program}{chr "PMA", "PMA", "PMA", "PMA", "PMA", "PMA", "PMA…}
\item{table_title}{chr "PMA Original and Panel Track Supplements – Ac…}
\item{metric_type}{chr "integer", "integer", "integer", "integer", "i…}
\item{performance_metric}{chr "Number Received", "Number Received", "Number
…}
\item{fy}{chr "2013", "2014", "2015", "2016", "2017", "2013"…}
\item{value}{chr "35", "43", "72", "72", "68", "1", "0", "1", "…}
}
}
\source{
\href{https://www.fda.gov/industry/medical-device-user-fee-amendments-mdufa/mdufa-reports}{FDA MDUFA Reports} # nolint: line_length_linter.
accessed 2023-11-29.
}
\usage{
mdufa3
}
\description{
FDA's MDUFA III Quarterly Medical Device User Performance Metrics in a
tidy data format.
}
\keyword{datasets}
