% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdufa4.R
\docType{data}
\name{mdufa4}
\alias{mdufa4}
\title{MDUFA IV}
\format{
A tibble with 15550 rows and 14 fields:

\describe{
\item{report_description}{chr "November 16, 2023 MDUFA IV Performance
Report…}
\item{report_link}{chr "https://www.fda.gov/media/173924/download?att…}
\item{report_date}{date 2023-11-16, 2023-11-16, 2023-11-16, 2023-11-1…}
\item{report_mdufa_period}{chr "MDUFA IV", "MDUFA IV", "MDUFA IV", "MDUFA
IV"…}
\item{source}{chr "Table 1.1 CDRH - PMA Original and Panel-Track…}
\item{page}{chr "17", "17", "17", "17", "17", "17", "17", "17"…}
\item{table_number}{chr "1.1", "1.1", "1.1", "1.1", "1.1", "1.1", "1.1…}
\item{organization}{chr "CDRH", "CDRH", "CDRH", "CDRH", "CDRH", "CDRH"…}
\item{program}{chr "PMA Original and Panel-Track Supplements", "P…}
\item{table_title}{chr "PMA Original and Panel-Track Supplements - Ac…}
\item{metric_type}{chr "integer", "integer", "integer", "integer", "i…}
\item{performance_metric}{chr "Number Received", "Number Received", "Number
…}
\item{fy}{chr "2018", "2019", "2020", "2021", "2022", "2018"…}
\item{value}{chr "74", "56", "77", "78", "45", "0", "0", "0", "…}
}
}
\source{
\href{https://www.fda.gov/industry/medical-device-user-fee-amendments-mdufa/mdufa-reports}{FDA MDUFA Reports} # nolint: line_length_linter.
accessed 2024-08-24.
}
\usage{
mdufa4
}
\description{
FDA's Quarterly Medical Device User Performance Metrics in a tidy data
format.
}
\keyword{datasets}
