% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdufa5.R
\docType{data}
\name{mdufa5}
\alias{mdufa5}
\title{MDUFA V}
\format{
A tibble with 16410 rows and 14 fields:

\describe{
\item{report_description}{chr "May 19, 2026 MDUFA V Performance Report",
"Ma…}
\item{report_link}{chr "https://www.fda.gov/media/192669/download?att…}
\item{report_date}{date 2026-05-19, 2026-05-19, 2026-05-19, 2026-05-1…}
\item{report_mdufa_period}{chr "MDUFA V", "MDUFA V", "MDUFA V", "MDUFA V",
"M…}
\item{source}{chr "Table 1.1 CDRH - PMA Original and Panel-Track…}
\item{page}{chr "33", "33", "33", "33", "33", "33", "33", "33"…}
\item{table_number}{chr "1.1", "1.1", "1.1", "1.1", "1.1", "1.1", "1.1…}
\item{organization}{chr "CDRH", "CDRH", "CDRH", "CDRH", "CDRH", "CDRH"…}
\item{program}{chr "PMA Original and Panel-Track Supplements", "P…}
\item{table_title}{chr "PMA Original and Panel-Track Supplements - Ac…}
\item{metric_type}{chr "integer", "integer", "integer", "integer", "i…}
\item{performance_metric}{chr "Number Received", "Number Received", "Number
…}
\item{fy}{chr "2023", "2024", "2025", "2026", "2027", "2023"…}
\item{value}{chr "73", "69", "89", "45", NA, "0", "0", "1", "1"…}
}
}
\source{
\href{https://www.fda.gov/industry/medical-device-user-fee-amendments-mdufa-fees/mdufa-reports}{FDA MDUFA Reports} # nolint: line_length_linter.
accessed 2026-06-01.
}
\usage{
mdufa5
}
\description{
FDA's Quarterly Medical Device User Performance Metrics in a tidy data
format.
}
\keyword{datasets}
