% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric_types.R
\name{metric_types}
\alias{metric_types}
\title{Match Performance Metrics to Data Types}
\usage{
metric_types(report_mdufa_period = "MDUFA IV")
}
\arguments{
\item{report_mdufa_period}{Text string defining whether the data is coming
from a MDUFA III, MDUFA IV, or MDUFA V report. Currently defaults to
\code{"MDUFA IV"}.}
}
\value{
A table with common performance metrics in the first column and
metric types (e.g. integer, percent, etc.) in the second column.
}
\description{
Match Performance Metrics to Data Types
}
