% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{oht_names}
\alias{oht_names}
\title{OHT Full Names}
\format{
Named character vector where names are OHT codes and values are
full names
}
\usage{
oht_names
}
\description{
Full names for OHT organizations as they appear in MDUFA IV and V reports.
Used for labeling plots and tables.
}
\keyword{datasets}
