% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_tables.R
\name{process_page}
\alias{process_page}
\title{Process a PDF Page}
\usage{
process_page(page_string)
}
\arguments{
\item{page_string}{The character string of text on the page.}
}
\value{
A list of the page's tables.
}
\description{
Process a PDF Page
}
