% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_tables.R
\name{process_page_m5}
\alias{process_page_m5}
\title{Process a PDF Page}
\usage{
process_page_m5(page_string, page_number)
}
\arguments{
\item{page_string}{The character string of text on the page.}

\item{page_number}{The page number based on the pdf index.}
}
\value{
A list of the page's tables.
}
\description{
These reports don't always have page numbers, complicating the procedure.
}
