% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_tables.R
\name{remove_unnecessary_text}
\alias{remove_unnecessary_text}
\title{Remove Unnecessary Text}
\usage{
remove_unnecessary_text(text_string)
}
\arguments{
\item{text_string}{The string to clean}
}
\value{
A string with anything matching \code{unnecessary_text_pattern()}
removed.
}
\description{
Remove unnecessary text, like section headings and asterisked notes about
RTA process implementation, from a text string. Uses the patterns defined in
\code{unnecessary_text_pattern()}.
}
