% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_oht.R
\name{rename_divisions_to_oht}
\alias{rename_divisions_to_oht}
\title{Rename Division to OHT (1:1 Mappings)}
\usage{
rename_divisions_to_oht(data)
}
\arguments{
\item{data}{MDUFA III dataset}
}
\value{
Dataset with renamed organizations
}
\description{
Renames divisions that have a direct 1:1 mapping to OHT organizations.
DCD -> OHT2, DNPMD -> OHT5, DOD -> OHT6, DRH -> OHT8.
}
\examples{
library(dplyr)
mdufa3 |>
  filter(organization == "DCD") |>
  rename_divisions_to_oht() |>
  pull(organization) |>
  unique()
}
