% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{report_dates_cols}
\alias{report_dates_cols}
\title{Expected Columns for report_dates Dataset}
\format{
Character vector of 5 column names
}
\usage{
report_dates_cols
}
\description{
Column names in the expected order for the report_dates dataset.
Contains report publication dates and their corresponding data cutoff dates.
}
\keyword{datasets}
