% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regex_patterns.R
\name{si_metric_pattern}
\alias{si_metric_pattern}
\title{Substantive Interaction Metric Pattern}
\usage{
si_metric_pattern()
}
\value{
a regular expression pattern
}
\description{
This function generates a regular expression pattern for matching performance
metrics associated with Substantive Interactions in text.
}
\details{
This function returns a regular expression pattern that matches three
different types of Substantive Interaction metrics: Average Number of
FDA/Industry days to Substantive Interaction, Xth Percentile FDA/Industry
days to Substantive Interaction, and Maximum FDA/Industry days to
Substantive Interaction. The function ignores case and does not match
across multiple lines.

The final word needs to be followed by a | for this expression to match.
}
\examples{
testthat::expect_true(
  stringr::str_detect(
    string = "\n40th Percentile FDA days to Substantive Interaction|",
    pattern = si_metric_pattern()
  )
)

testthat::expect_true(
  stringr::str_detect(
    string = "\nMaximum FDA days to\nSubstantive Interaction|",
    pattern = si_metric_pattern()
  )
)

testthat::expect_true(
  stringr::str_detect(
    string = "\nAverage number of FDA days to\nSubstantive Interaction|",
    pattern = si_metric_pattern()
  )
)
}
