% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regex_patterns.R
\name{table_title_text_pattern}
\alias{table_title_text_pattern}
\title{Generate a regular expression pattern for table titles}
\usage{
table_title_text_pattern()
}
\value{
A regular expression pattern
}
\description{
This function generates a regular expression pattern for matching table
titles in text. The pattern matches a string that starts with "Table",
followed by a number, a period, and another number, followed by a space and
any characters that are not a vertical bar (|).
}
