% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type_metrics.R
\name{type_metrics}
\alias{type_metrics}
\title{Type Metrics}
\usage{
type_metrics(data)
}
\arguments{
\item{data}{A performance table, such as \code{mdufa::quarterly_performance}.}
}
\value{
A performance table with a list column for \code{value}.
}
\description{
Return a performance table with a list column for \code{value} so that
multiple metric types can be accommodated.
}
\examples{
type_metrics(mdufa::mdufa4)
}
