% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regex_patterns.R
\name{unnecessary_text_pattern}
\alias{unnecessary_text_pattern}
\title{Unnecessary Text Pattern}
\usage{
unnecessary_text_pattern()
}
\value{
A regex pattern that defines unnecessary text that can be removed.
}
\description{
Unnecessary Text Pattern
}
