% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regex_patterns.R
\name{within_days_pattern}
\alias{within_days_pattern}
\title{Generate a regular expression pattern for within days}
\usage{
within_days_pattern()
}
\value{
A regular expression pattern
}
\description{
This function generates a regular expression pattern for matching within days
in text. The pattern matches a string that starts with "Within", followed by
a space and one or more digits.
}
