% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_clean_tibble.R
\name{make_clean_tibble}
\alias{make_clean_tibble}
\title{Make a Tibble with Correct Names}
\usage{
make_clean_tibble(data)
}
\arguments{
\item{data}{A tibble}
}
\value{
A tibble with names that are harmonized with other tables from the
report.
}
\description{
Make a Tibble with Correct Names
}
