% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_dates.R
\docType{data}
\name{report_dates}
\alias{report_dates}
\title{Report Dates}
\format{
A tibble with 82 rows and 5 fields:

\describe{
\item{report_date}{date 2009-01-28, 2009-05-04, 2009-08-04, 2009-11-1…}
\item{report_cutoff_date}{date 2008-12-31, 2009-03-31, 2009-06-30,
2009-09-3…}
\item{report_mdufa_period}{chr "MDUFA II", "MDUFA II", "MDUFA II", "MDUFA
II"…}
\item{report_description}{chr "January 28, 2009 FDA Performance Report
(Arch…}
\item{report_link}{chr "https://wayback.archive-it.org/7993/201701120…}
}
}
\source{
\href{https://www.fda.gov/industry/medical-device-user-fee-amendments-mdufa-fees/mdufa-reports}{FDA MDUFA Reports} # nolint: line_length_linter.
accessed 2026-02-25.
}
\usage{
report_dates
}
\description{
Mapping of MDUFA quarterly report publication dates to their data cutoff
dates. Each report contains performance data through a specific cutoff
date (typically the end of a fiscal quarter).
}
\keyword{datasets}
